set bar_start_tick when running as timebase master
git-svn-id: svn+ssh://kobol/srv/svn/klick/trunk@247 c380f6eb-e629-0410-afc2-bdd98b771d42

+++ src/metronome_map.hh
@@ -48,7 +48,7 @@ class MetronomeMap
     virtual void timebase_callback(position_t *);
 
   private:
-    static double const TICKS_PER_BEAT = 1920.0;
+    static int const TICKS_PER_BEAT = 1920;
 
     // transport position
     nframes_t _current;
